+++ caching-service/src/main/java/org/zowe/apiml/caching/service/inmemory/InMemoryStorage.java
@@ -127,7 +127,7 @@ public class InMemoryStorage implements Storage {
     }
 
     @Override
-    public void deleteItemFromMap(String serviceId, String mapKey, String itemKey) {
+    public void deleteItemFromMap(String serviceId, String mapKey) {
         throw new StorageException(Messages.INCOMPATIBLE_STORAGE_METHOD.getKey(), Messages.INCOMPATIBLE_STORAGE_METHOD.getStatus());
     }
