+++ caching-service/src/main/java/org/zowe/apiml/caching/service/vsam/VsamStorage.java
@@ -210,7 +210,7 @@ public class VsamStorage implements Storage {
     }
 
     @Override
-    public void deleteItemFromMap(String serviceId, String mapKey, String itemKey) {
+    public void deleteItemFromMap(String serviceId, String mapKey) {
         throw new StorageException(Messages.INCOMPATIBLE_STORAGE_METHOD.getKey(), Messages.INCOMPATIBLE_STORAGE_METHOD.getStatus());
     }
 }
